HADOOP-18197. Upgrade protobuf to 3.20.1
This patch bumps up the protobuf version so that Hadoop
is not a vulnerable to CVE-2021-22569.

I'm not renaming the module hadoop-shaded-protobuf_3_7
because that significantly complicates imports/upgrading.
That said, I don't see why the version number needed to be
included there. We will have to live with that.

This also fixes up the parent POM references in the child modules
as IntelliJ requires a full path.

Testing: needs to go through hadoop built with the updated jar and
with its own protobuf version marker updated.
Verified hadoop compiles on a macbook m1.